almost done with audit feature

+++ ProjectForms/CommentManagementForm.cs
@@ -1,4 +1,4 @@
-﻿using ProjectManagementBusinessObjects;
+using ProjectManagement.Model;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
@@ -15,9 +15,9 @@ namespace ProjectForms
     public partial class CommentManagementForm : Form
     {
         private ProjectManagementDBContext context;
-        private ProjectManagementBusinessObjects.Task selectedTask;
+        private ProjectManagement.Model.Task selectedTask;
 
-        public CommentManagementForm(ProjectManagementBusinessObjects.Task task)
+        public CommentManagementForm(ProjectManagement.Model.Task task)
         {
             InitializeComponent();
             context = new ProjectManagementDBContext();
@@ -42,8 +42,6 @@ namespace ProjectForms
                     Comment_Text = i.CommentText,
                     UserID = i.UserId,
                     Task_ID = i.TaskId,
-
-
                 }).ToList();
 
             dgvComments.DataSource = comments;
@@ -51,53 +49,69 @@ namespace ProjectForms
 
         private void btnAddComment_Click(object sender, EventArgs e)
         {
-            try
+
+
+            int userid = Convert.ToInt32(context.Users.Where(x => x.Email == Global.SelectedUser.Email).FirstOrDefault()?.UserId);
+            if (string.IsNullOrWhiteSpace(txtComment.Text))
             {
-                if (string.IsNullOrWhiteSpace(txtComment.Text))
-                {
-                    MessageBox.Show("Please enter a comment.", "Validation Error", MessageBoxButtons.OK, MessageBoxIcon.Warning);
-                    return;
-                }
-                byte[] currentDateBytes = BitConverter.GetBytes(DateTime.Now.Ticks);
-                Comment newComment = new Comment
-                {
-                    CommentText = txtComment.Text,
-                    CommentTimestamp = currentDateBytes,
-                    UserId = Convert.ToInt32(context.Users.Where(x => x.Email == Global.SelectedUser.Email).FirstOrDefault()?.UserId),
-                    TaskId = selectedTask.TaskId
-                };
-
-                context.Comments.Add(newComment);
-                context.SaveChanges();
-
-                LoadComments();
-                txtComment.Clear();
-
+                MessageBox.Show("Please enter a comment.", "Validation Error", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                return;
             }
-            catch(Exception ex)
+
+
+            Comment newComment = new Comment
             {
-                MessageBox.Show($"An error occurred: {ex.Message}");
-                int id = context.Users.Where(x => x.Email == Global.SelectedUser.Email).FirstOrDefault().UserId;
-                LoggingService log = new LoggingService(context);
-                log.LogException(ex, id);
-            }
-           
+                CommentText = txtComment.Text,
+
+                UserId = Convert.ToInt32(context.Users.Where(x => x.Email == Global.SelectedUser.Email).FirstOrDefault()?.UserId),
+                TaskId = selectedTask.TaskId
+            };
+
+            var auditLog = new Audit
+            {
+                ChangeType = "Create",
+                TableName = "Comment",
+                RecordId = newComment.CommentId,
+                CurrentValue = GetCommentValues(newComment),
+                OldValue = null,
+                UserId = userid,
+            };
+
+            context.Set<ProjectManagement.Model.Audit>().Add(auditLog);
+            context.Comments.Add(newComment);
+            context.SaveChanges();
+
+            MessageBox.Show("Comment added successfully!", "Success", MessageBoxButtons.OK);
+
+            LoadComments();
+            txtComment.Clear();
+        
         }
+
 
         private void btnDelete_Click(object sender, EventArgs e)
         {
 
-            try
+            if (dgvComments.SelectedCells.Count > 0)
             {
+                int userid = Convert.ToInt32(context.Users.Where(x => x.Email == Global.SelectedUser.Email).FirstOrDefault()?.UserId);
                 int commentId = Convert.ToInt32(dgvComments.SelectedCells[0].OwningRow.Cells[0].Value);
                 Comment comment = context.Comments.Find(commentId);
-
-
                 if (comment != null)
                 {
                     DialogResult result = MessageBox.Show("Are you sure you want to delete this comment?", "Confirmation", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
                     if (result == DialogResult.Yes)
                     {
+                        var auditLog = new Audit
+                        {
+                            ChangeType = "Delete",
+                            TableName = "Comment",
+                            RecordId = commentId,
+                            OldValue = GetCommentValues(comment),
+                            UserId = userid,
+                        };
+                       
+                        context.Set<ProjectManagement.Model.Audit>().Add(auditLog);
                         context.Comments.Remove(comment);
                         context.SaveChanges();
                         LoadComments();
@@ -109,23 +123,16 @@ namespace ProjectForms
                     MessageBox.Show("Unable to find the selected comment.");
                 }
             }
-            catch (Exception ex)
+            else
             {
-                MessageBox.Show($"An error occurred: {ex.Message}");
-                int id = context.Users.Where(x => x.Email == Global.SelectedUser.Email).FirstOrDefault().UserId;
-                LoggingService log = new LoggingService(context);
-                log.LogException(ex, id);
+                MessageBox.Show("Please select a comment.");
             }
-           
-
-
-
-
         }
 
         private void btnEdit_Click(object sender, EventArgs e)
         {
-            try
+
+            if (dgvComments.SelectedCells.Count > 0)
             {
                 int commentId = Convert.ToInt32(dgvComments.SelectedCells[0].OwningRow.Cells[0].Value);
                 Comment comment = context.Comments.Find(commentId);
@@ -134,22 +141,16 @@ namespace ProjectForms
                 {
                     EditCommentForm editForm = new EditCommentForm(comment, context);
                     DialogResult result = editForm.ShowDialog();
-
-
                 }
                 else
                 {
                     MessageBox.Show("Unable to find the selected comment.");
                 }
             }
-            catch (Exception ex)
+            else
             {
-                MessageBox.Show($"An error occurred: {ex.Message}");
-                int id = context.Users.Where(x => x.Email == Global.SelectedUser.Email).FirstOrDefault().UserId;
-                LoggingService log = new LoggingService(context);
-                log.LogException(ex, id);
+                MessageBox.Show("Please select a comment.");
             }
-           
 
         }
 
@@ -160,6 +161,10 @@ namespace ProjectForms
             manageForm.Show();
         }
 
+        private string GetCommentValues(Comment comment)
+        {
+            return $"CommentId: {comment.CommentId}, TaskId: {comment.TaskId}, Text: {comment.CommentText}";
+        }
 
     }
 }
